+++ packages/platform-server/router/index.ts
@@ -160,4 +160,3 @@ export function NvModuleFactoryLoader(loadChild: Function, currentUrlPath: strin
 
   return loadModuleInstance;
 }
-